moved visitor classes into visitor namespaces
