+++ src/JMS/Serializer/Handler/BasicHandler.php
@@ -22,7 +22,7 @@ use JMS\Serializer\Context;
 use JMS\Serializer\Exception\RuntimeException;
 use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\TypeDefinition;
-use JMS\Serializer\VisitorInterface;
+use JMS\Serializer\Visitor\VisitorInterface;
 
 class BasicHandler implements SubscribingHandlerInterface
 {
