+++ src/JMS/Serializer/Handler/FormErrorHandler.php
@@ -18,12 +18,12 @@
 
 namespace JMS\Serializer\Handler;
 
-use JMS\Serializer\GenericSerializationVisitor;
+use JMS\Serializer\Visitor\GenericSerializationVisitor;
 use JMS\Serializer\Graph\GraphNavigator;
-use JMS\Serializer\JsonSerializationVisitor;
+use JMS\Serializer\Visitor\JsonSerializationVisitor;
 use JMS\Serializer\TypeDefinition;
-use JMS\Serializer\XmlSerializationVisitor;
-use JMS\Serializer\YamlSerializationVisitor;
+use JMS\Serializer\Visitor\XmlSerializationVisitor;
+use JMS\Serializer\Visitor\YamlSerializationVisitor;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Translation\TranslatorInterface;
