+++ src/JMS/Serializer/Visitor/GenericSerializationVisitor.php
@@ -16,12 +16,14 @@
  * limitations under the License.
  */
 
-namespace JMS\Serializer;
+namespace JMS\Serializer\Visitor;
 
+use JMS\Serializer\Context;
 use JMS\Serializer\Exception\InvalidArgumentException;
 use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\Metadata\ClassMetadata;
 use JMS\Serializer\Metadata\PropertyMetadata;
+use JMS\Serializer\TypeDefinition;
 
 abstract class GenericSerializationVisitor extends AbstractVisitor
 {
