+++ src/JMS/Serializer/Visitor/VisitorInterface.php
@@ -16,11 +16,13 @@
  * limitations under the License.
  */
 
-namespace JMS\Serializer;
+namespace JMS\Serializer\Visitor;
 
+use JMS\Serializer\Context;
 use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\Metadata\ClassMetadata;
 use JMS\Serializer\Metadata\PropertyMetadata;
+use JMS\Serializer\TypeDefinition;
 
 /**
  * Interface for visitors.
