+++ src/JMS/Serializer/Visitor/YamlSerializationVisitor.php
@@ -16,12 +16,14 @@
  * limitations under the License.
  */
 
-namespace JMS\Serializer;
+namespace JMS\Serializer\Visitor;
 
+use JMS\Serializer\Context;
 use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\Metadata\ClassMetadata;
 use JMS\Serializer\Metadata\PropertyMetadata;
 use JMS\Serializer\Naming\PropertyNamingStrategyInterface;
+use JMS\Serializer\TypeDefinition;
 use JMS\Serializer\Util\Writer;
 use Symfony\Component\Yaml\Inline;
